Move tables and rows to a partial class
