+++ src/Compilers/Core/Portable/PEWriter/MetadataWriter.PortablePdb.cs
@@ -11,71 +11,6 @@ namespace Microsoft.Cci
 {
     partial class MetadataWriter
     {
-        private struct DocumentRow
-        {
-            public BlobIdx Name;          
-            public uint HashAlgorithm; // Guid
-            public BlobIdx Hash;       
-            public uint Language;      // Guid
-        }
-
-        private struct MethodDebugInformationRow
-        {
-            public uint Document;       // DocumentRid
-            public BlobIdx SequencePoints;
-        }
-
-        private struct LocalScopeRow
-        {
-            public uint Method;         // MethodRid
-            public uint ImportScope;    // ImportScopeRid
-            public uint VariableList;   // LocalVariableRid
-            public uint ConstantList;   // LocalConstantRid
-            public uint StartOffset;  
-            public uint Length;                  
-        }
-
-        private struct LocalVariableRow
-        {
-            public ushort Attributes;
-            public ushort Index;
-            public StringIdx Name;     
-        }
-
-        private struct LocalConstantRow
-        {
-            public StringIdx Name;     
-            public BlobIdx Signature;
-        }
-
-        private struct ImportScopeRow
-        {
-            public uint Parent;        // ImportScopeRid
-            public BlobIdx Imports;   
-        }
-
-        private struct StateMachineMethodRow
-        {
-            public uint MoveNextMethod;      // MethodRid
-            public uint KickoffMethod;       // MethodRid
-        }
-
-        private struct CustomDebugInformationRow
-        {
-            public uint Parent;       // HasCustomDebugInformation coded index
-            public uint Kind;         // Guid
-            public BlobIdx Value;     
-        }
-
-        private readonly List<DocumentRow> _documentTable = new List<DocumentRow>();
-        private readonly List<MethodDebugInformationRow> _methodDebugInformationTable = new List<MethodDebugInformationRow>();
-        private readonly List<LocalScopeRow> _localScopeTable = new List<LocalScopeRow>();
-        private readonly List<LocalVariableRow> _localVariableTable = new List<LocalVariableRow>();
-        private readonly List<LocalConstantRow> _localConstantTable = new List<LocalConstantRow>();
-        private readonly List<ImportScopeRow> _importScopeTable = new List<ImportScopeRow>();
-        private readonly List<StateMachineMethodRow> _stateMachineMethodTable = new List<StateMachineMethodRow>();
-        private readonly List<CustomDebugInformationRow> _customDebugInformationTable = new List<CustomDebugInformationRow>();
-
         private readonly Dictionary<DebugSourceDocument, int> _documentIndex = new Dictionary<DebugSourceDocument, int>();
         private readonly Dictionary<IImportScope, int> _scopeIndex = new Dictionary<IImportScope, int>();
